Using Optimizations breakdown from Koku-ui microfrontend
